Fix overflows when finding similiar colors

+++ src/color.rs
@@ -36,9 +36,9 @@ impl Color {
     pub fn similiar(&self) -> Color {
         let Color { r, g, b, a } = *self;
 
-        let new_r = get_value_around(r, MAX_SIMILIARITY_OFFSET).clamp(0, 255);
-        let new_g = get_value_around(g, MAX_SIMILIARITY_OFFSET).clamp(0, 255);
-        let new_b = get_value_around(b, MAX_SIMILIARITY_OFFSET).clamp(0, 255);
+        let new_r = get_value_around(r as i32, MAX_SIMILIARITY_OFFSET as i32).clamp(0, 255) as u8;
+        let new_g = get_value_around(g as i32, MAX_SIMILIARITY_OFFSET as i32).clamp(0, 255) as u8;
+        let new_b = get_value_around(b as i32, MAX_SIMILIARITY_OFFSET as i32).clamp(0, 255) as u8;
 
         Color::rgba(new_r, new_g, new_b, a)
     }
